refactor: :recycle: modularize common names + synonyms
